requirements.full.txt
# Minimal runtime
numpy==1.24.3
huggingface-hub==0.30.2
easydict==1.13
pandas==2.2.2
einops==0.7.0
opencv-python==4.11.0.86
matplotlib==3.8.3
seaborn==0.13.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
pypng==0.20220715.0
kornia==0.7.3
flow-vis==0.1
moviepy==1.0.3
mediapy==1.2.0
rerun-sdk==0.21.0

# Training / baselines
torchdata==0.11.0
lightning==2.4.0
timm==0.6.7
prettytable==3.10.0
# tensorflow==2.12.1
# tensorflow-datasets==4.9.8
# tensorflow-graphics==2021.12.3
tensorboard==2.12.3
tqdm==4.67.1
gpustat==1.1.1
hydra-core==1.3.2
wandb==0.19.9
rich==14.0.0
